requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
websockets>=12.0
requests>=2.31.0
pyperclip>=1.8.2
pyautogui>=0.9.54
keyboard>=0.13.5
pystray>=0.19.5
Pillow>=10.0.0